+++ src/parsing_snowflake.jl
@@ -55,6 +55,7 @@ function expr_to_sql_snowflake(expr, sq; from_summarize::Bool)
                window_clause = construct_window_clause(sq, from_cumsum = true)
                return  "SUM($(string(a))) $(window_clause)"
             end
+
         #stats agg
         elseif @capture(x, std(a_))
             if from_summarize